requirements.txt
aiohttp==3.7.3
aniso8601==7.0.0
asgiref==3.3.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
cffi==1.14.4
channels==3.0.3
chardet==3.0.4
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.3
django-background-tasks==1.2.5
django-channels-graphql-ws==0.7.5
django-compat==1.0.15
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.2
graphene==2.1.8
graphene-django==2.13.0
graphene-django-extras==0.4.9
graphql-core==2.3.2
graphql-relay==2.0.1
hyperlink==20.0.1
idna==3.0
incremental==17.5.0
install==1.3.4
msgpack==1.0.2
multidict==5.1.0
promise==2.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2020.4
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
Unidecode==1.1.1
whitenoise==5.2.0
yarl==1.6.3
zope.interface==5.2.0
